test: extend tests
