+++ runner/test-data/collectSteps/Example.feature.md
@@ -3,3 +3,5 @@
 ## Calculator
 
 Given I add 1 and 1
+
+And I subtract 2 from 3
